complete repo update
:wq#	new file:   test/run_bench_all.sh

+++ example/main.c
@@ -13,9 +13,9 @@ static void fill_spd_matrix(double *c, int n)
 {
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
-            c[n * i + j] = corr(i, j, n);
+            c[(long)n * i + j] = corr(i, j, n);
         }
-        c[n * i + i] = 1.0;
+        c[(long)n * i + i] = 1.0;
     }
 }
 
@@ -24,7 +24,7 @@ static double log_determinant(const double *c, int n)
 {
     double logdet = 0.0;
     for (int p = 0; p < n; p++) {
-        logdet += log(c[n * p + p]);
+        logdet += log(c[(long)n * p + p]);
     }
     return 2.0 * logdet;
 }
@@ -43,10 +43,10 @@ int main(int argc, char *argv[])
 
     printf("Cholesky factorization: n = %d\n", n);
 
-    double *c = (double *)malloc((size_t)n * n * sizeof(double));
+    double *c = (double *)malloc((size_t)n * (size_t)n * sizeof(double));
     if (!c) {
         fprintf(stderr, "Error: failed to allocate %zu bytes\n",
-                (size_t)n * n * sizeof(double));
+                (size_t)n * (size_t)n * sizeof(double));
         return 1;
     }
